+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `<Table />` - additional download options to reflect filtered rows.
 - `<Table />` - show dot badge for download when options are allowed.
+- `<Table />` - localStorage based caching mechanism to remember user preference of column visibility and sort by.
 
 ## [1.1.3] - 2020-06-29
 ### Fixed