+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2.0] - 2020-06-30
 ### Added
 - `<Table />` - additional download options to reflect filtered rows.
 - `<Table />` - show dot badge for download when options are allowed.